+++ lib/domain/record/components/add_pill_sheet_group/add_pill_sheet_group_empty_frame.dart
@@ -1,6 +1,6 @@
 import 'package:pilll/analytics.dart';
 import 'package:pilll/domain/record/components/add_pill_sheet_group/add_pill_sheet_group_page.dart';
-import 'package:pilll/domain/record/record_page_store.dart';
+import 'package:pilll/domain/record/record_page_state_notifier.dart';
 import 'package:pilll/entity/setting.codegen.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
