+++ lib/domain/record/components/supports/components/display_number_setting/store.dart
@@ -1,5 +1,5 @@
 import 'package:pilll/database/batch.dart';
-import 'package:pilll/domain/record/record_page_store.dart';
+import 'package:pilll/domain/record/record_page_state_notifier.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
