+++ test/domain/record/record_page_button_state_test.dart
@@ -2,7 +2,7 @@ import 'package:pilll/analytics.dart';
 import 'package:pilll/domain/record/components/button/cancel_button.dart';
 import 'package:pilll/domain/record/components/button/record_page_button.dart';
 import 'package:pilll/domain/record/components/button/taken_button.dart';
-import 'package:pilll/domain/record/record_page_store.dart';
+import 'package:pilll/domain/record/record_page_state_notifier.dart';
 import 'package:pilll/entity/pill_sheet.codegen.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/entity/pill_sheet_type.dart';
